# rest/api/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Iniate a websocket connection",
        "title": "blockmania API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/api/pubsub/ws": {
            "get": {
                "description": "Iniate a websocket connection",
                "tags": [
                    "pubsub"
                ],
                "summary": "Initiate a websocket connection in order to subscribed to object saved in chainspace by the current node",
                "operationId": "websocket",
                "responses": {
                    "204": {},
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/transaction": {
            "post": {
                "description": "Add a new transction in the blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blockmania"
                ],
                "summary": "Add a new transaction in the blockchain",
                "operationId": "addTransction",
                "parameters": [
                    {
                        "description": "transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.TransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/transactions": {
            "post": {
                "description": "Add a list of new transctions in the blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blockmania"
                ],
                "summary": "Add a list of new transactions in the blockchain",
                "operationId": "addTransactions",
                "parameters": [
                    {
                        "description": "transactions",
                        "name": "transactions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.TransactionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "api.TransactionRequest": {
            "type": "object",
            "properties": {
                "tx": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.TransactionsRequest": {
            "type": "object",
            "properties": {
                "txs": {
                    "type": "array",
                    "items": {
                        "type": "\u0026{%!s(token.Pos=410) \u003cnil\u003e byte}"
                    }
                }
            }
        }
    }
}